+++ quiltcore/__init__.py
@@ -24,13 +24,13 @@ manages the 'quiltcore.yaml' configuration file.
 
 """
 
-from .builder2 import FolderBuilder  # noqa: F401
+from .builder import FolderBuilder  # noqa: F401
 from .domain import Domain  # noqa: F401
-from .entry2 import Entry2  # noqa: F401
+from .entry import Entry2  # noqa: F401
 from .factory import Factory, quilt  # noqa: F401
 from .udg.header import Header  # noqa: F401
-from .manifest2 import Manifest2  # noqa: F401
-from .namespace2 import Namespace2  # noqa: F401
+from .manifest import Manifest2  # noqa: F401
+from .namespace import Namespace2  # noqa: F401
 from .scheme import Scheme  # noqa: F401
 from .table3 import Table3  # noqa: F401
 from .table4 import Table4  # noqa: F401
