+++ quiltcore/manifest.py
