+++ tests/test_manifest.py
